Add pretty exception message printing

+++ src/main/java/com/codecool/wineREST/controllers/ProducentController.java
@@ -4,7 +4,8 @@ import com.codecool.wineREST.entities.Producent;
 import com.codecool.wineREST.exceptions.FkViolationException;
 import com.codecool.wineREST.services.ProducentArchiveService;
 import com.codecool.wineREST.services.ProducentService;
-import org.hibernate.exception.ConstraintViolationException;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
@@ -16,10 +17,11 @@ import java.util.NoSuchElementException;
 @RequestMapping("/producents")
 public class ProducentController {
 
+    private Logger logger = LogManager.getLogger();
     @Autowired
-    ProducentService producentService;
+    private ProducentService producentService;
     @Autowired
-    ProducentArchiveService producentArchiveService;
+    private ProducentArchiveService producentArchiveService;
 
     @RequestMapping(path = "/{idProducent}", method = RequestMethod.GET)
     public Producent getById(@PathVariable(value = "idProducent") long idProducent) {
@@ -30,15 +32,27 @@ public class ProducentController {
     @ResponseStatus(HttpStatus.CREATED)
     public void createProducent(@RequestBody Producent producent) {
         this.producentService.createProducent(producent.getName());
+        logger.info("Created new region");
     }
 
     @RequestMapping(path = "/{idProducent}", method = RequestMethod.DELETE)
-    public void archiveProducent(@PathVariable(value = "idProducent") long idProducent) throws FkViolationException {
+    public void archiveProducent(@PathVariable(value = "idProducent") long idProducent) throws NoSuchElementException,
+            FkViolationException {
         Producent producent = this.producentService.findById(idProducent);
+        if(producent == null) {
+            throw new NoSuchElementException("There is no producent with id: " + idProducent);
+        }
         try {
             this.producentService.deleteProducent(producent);
             this.producentArchiveService.archive(producent);
         } catch (DataIntegrityViolationException e) {
+            logger.error("Constraint violation: prodcent of id " + producent.getIdProducent() + " is attached to wine" +
+                    " existing in the DB");
+            logger.debug("This is a debug message");
+            logger.info("This is an info message");
+            logger.warn("This is a warn message");
+            logger.error("This is an error message");
+            logger.fatal("This is a fatal message");
             throw new FkViolationException("This producent is attached to wine existing in the DB");
         }
     }
@@ -58,5 +72,12 @@ public class ProducentController {
     @ExceptionHandler(FkViolationException.class)
     public String return400(FkViolationException e) {
         return e.getMessage();
+    }
+
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    @ExceptionHandler(NoSuchElementException.class)
+    public String return400(NoSuchElementException ex) {
+        return ex.getMessage();
+
     }
 }
